+++ .meteor/packages
@@ -16,7 +16,6 @@ es5-shim@4.8.0                # ECMAScript 5 compatibility for older browsers
 ecmascript@0.12.0              # Enable ECMAScript2015+ syntax in app code
 shell-server@0.4.0            # Server-side component of the `meteor shell` command
 
-kadira:flow-router
 react-meteor-data
 accounts-password@1.5.1
 nimble:restivus
@@ -27,4 +26,3 @@ meteortesting:mocha
 ostrio:files
 underscore@1.0.10
 alanning:roles
-twbs:bootstrap
