+++ .meteor/versions
@@ -41,7 +41,6 @@ id-map@1.1.0
 inter-process-messaging@0.1.0
 jquery@1.11.11
 juliancwirko:postcss@2.0.3
-kadira:flow-router@2.12.1
 launch-screen@1.1.1
 less@2.8.0
 livedata@1.0.18
@@ -76,7 +75,6 @@ promise@0.11.1
 random@1.1.0
 rate-limit@1.0.9
 react-meteor-data@0.2.16
-reactive-dict@1.2.1
 reactive-var@1.0.11
 reload@1.2.0
 retry@1.1.0
@@ -96,7 +94,6 @@ templating-runtime@1.3.2
 templating-tools@1.1.2
 tmeasday:check-npm-versions@0.3.2
 tracker@1.2.0
-twbs:bootstrap@3.3.6
 ui@1.0.13
 underscore@1.0.10
 url@1.2.0
