+++ client/pages/index-page/index-page.jsx
@@ -1,0 +1,57 @@
+import React from 'react';
+import { connect } from 'react-redux';
+import { Meteor } from 'meteor/meteor';
+import { withTracker } from 'meteor/react-meteor-data';
+import { styles } from "./styles";
+import { LocaleAction } from "../../redux/actions/locale-action";
+
+class Component extends React.Component {
+
+	constructor(props) {
+		super(props);
+	}
+
+	render() {
+		return (
+			<React.Fragment>
+				<div
+					style={styles.page}
+				>
+					<h1>
+						{this.props.strings['welcome']}
+					</h1>
+					<p>
+						{this.props.strings['brief']}
+					</p>
+				</div>
+			</React.Fragment>
+		);
+	}
+
+	componentDidMount() {
+		this.props.dispatch(LocaleAction.set('en'));
+	}
+
+}
+
+const Tracker = withTracker(() => {
+	Meteor.subscribe('users_db');
+	return {
+		Meteor: {
+			collection: {
+				users: Meteor.users.find().fetch()
+			},
+			user: Meteor.user(),
+			userId: Meteor.userId(),
+			status: Meteor.status(),
+			loggingIn: Meteor.loggingIn()
+		}
+	};
+})(Component);
+
+export const IndexPage = connect((store) => {
+	return {
+		logs: store['LoggerReducer']['logs'],
+		strings: store['LocaleReducer']['strings']
+	};
+})(Tracker);
