+++ client/public/index.html
@@ -1,0 +1,3 @@
+<body>
+<div id="root"></div>
+</body>
