+++ package.json
@@ -7,16 +7,18 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.1.2",
+    "antd": "^3.10.8",
     "bcrypt": "^3.0.0",
     "bootstrap": "^4.1.1",
     "gridfs-stream": "^1.1.1",
+    "history": "^4.7.2",
     "meteor-node-stubs": "^0.3.2",
     "prop-types": "^15.6.2",
     "react": "^16.4.0",
-    "react-bootstrap": "^0.32.1",
     "react-dom": "^16.4.0",
     "react-mounter": "^1.2.0",
     "react-redux": "^5.0.7",
+    "react-router": "^4.3.1",
     "redux": "^4.0.0",
     "redux-saga": "^0.16.0"
   },
